+++ game/src/screens/match_screen/match_simulator/match_buffer/match_buffer_entries/match_buffer_teams_entry.gd
@@ -8,7 +8,14 @@ extends MatchBufferEntry
 var home_pos: Array[Vector2]
 var home_head_look: Array[Vector2]
 var home_info: Array[String]
+var home_skintone: Array[String]
+var home_haircolor: Array[String]
+var home_eyecolor: Array[String]
 
 var away_pos: Array[Vector2]
 var away_head_look: Array[Vector2]
 var away_info: Array[String]
+var away_skintone: Array[String]
+var away_haircolor: Array[String]
+var away_eyecolor: Array[String]
+