+++ game/src/screens/match_screen/match_simulator/visual_match/visual_match.gd
@@ -61,3 +61,4 @@ func show_actors() -> void:
 	ball.show()
 	home_team.show()
 	away_team.show()
+